+++ .github/workflows/ci.yml
@@ -102,6 +102,5 @@ jobs:
       - name: Check documentation
         run: |
           export RUSTDOCFLAGS="-D warnings"
-          cargo doc --no-deps --features serde-extensions,virt
-          cargo doc --no-deps --package trussed-core --all-features
+          cargo doc --no-deps --workspace --all-features
         if: matrix.target == 'x86_64-unknown-linux-gnu'
