+++ src/error.rs
@@ -1,0 +1,1 @@
+pub use trussed_core::{Error, Result};
