+++ src/interrupt.rs
@@ -1,0 +1,1 @@
+pub use trussed_core::{FromU8Error, InterruptFlag, InterruptState};
