+++ src/lib.rs
@@ -28,6 +28,8 @@ pub mod api;
 pub mod backend;
 pub mod client;
 pub mod config;
+pub mod error;
+pub mod interrupt;
 pub mod key;
 #[cfg(feature = "crypto-client")]
 pub mod mechanisms;
@@ -52,7 +54,7 @@ pub use error::Error;
 pub use platform::Platform;
 pub use service::Service;
 
-pub use trussed_core::{block, error, interrupt, syscall, try_syscall};
+pub use trussed_core::{block, syscall, try_syscall};
 
 pub use cbor_smol::cbor_deserialize;
 pub use heapless_bytes::Bytes;
